refactor: FlexBox.vueをatoms/へ
