+++ src/pages/index.vue
@@ -14,7 +14,7 @@ import ArticleCardGrid from "@/components/organisms/ArticleCardGrid.vue";
 import PageSet from "@/modules/PageSet";
 import SzppIcon from "@/components/atoms/SzppIcon.vue";
 import SimpleButton from "@/components/atoms/SimpleButton.vue";
-import FlexBox from "@/components/FlexBox.vue";
+import FlexBox from "@/components/atoms/FlexBox.vue";
 </script>
 
 <template>
